requirements.txt
pyarrow==14.0.2
turbodbc==4.5.10
adbc_driver_manager==0.10.0
adbc_driver_postgresql==0.10.0
adbc-driver-snowflake==0.10.0
numpy==1.26.4
pybind11==2.11.1
wheel==0.42.0
vertica_python==1.3.8
psycopg2==2.9.9
jaydebeapi==1.2.3
tabulate==0.9.0
PyYAML==6.0.1
python-dotenv==1.0.1
attrs>=21.4.0,<=23.1.0
cattrs>=22.1.0,<=23.2.3
